requirements.txt
# Core dependencies
streamlit>=1.28.0
pandas>=1.5.0
numpy>=1.24.0


sentence-transformers>=2.2.0
scikit-learn>=1.3.0
faiss-cpu>=1.7.4


mlflow>=2.8.0
databricks-cli>=0.18.0
databricks-sdk>=0.12.0


openai>=1.3.0


pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
python-dotenv>=1.0.0


psutil>=5.9.0